Condense test files, add filterTodosByBody() and corresponding condition
Moved all the filtering test files for todos into one test file while adding tests for filterTodosByBody().
